Separate Main component, standardize center width
add names to Contributor
