+++ src/variables.scss
@@ -35,3 +35,4 @@ $opacityColors: ();
 $borderRadius: 10px;
 $border: 1px solid opacity("secondaryBlack-30");
 $shadow: 0 8px 8px 2px color.adjust(#000, $alpha: -0.9);
+$centerProportion: 75%;
